requirements.txt
alembic==1.11.1
amqp==5.1.1
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.38
python-dateutil==2.8.2
python-slugify==8.0.1
six==1.16.0
SQLAlchemy==2.0.17
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
